+++ scripts.py
@@ -7,3 +7,13 @@ def test():
     `poetry run python -u -m unittest discover`
     """
     subprocess.run(["python", "-u", "-m", "unittest", "discover"])
+
+
+def parse_test():
+    """
+    Run xer parser unittests. Equivalent to:
+    `poetry run python -u -m unittest discover`
+    """
+    subprocess.run(
+        ["python", "-m", "unittest", "tests.test_parser.TestParser.test_create_xer"]
+    )